refa: 컨트롤러 하나로 합침
