+++ src/test/java/chess/domain/board/SquareTest.java
@@ -1,4 +1,4 @@
-package domain.board;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,6 +6,10 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.board.File;
+import domain.board.Rank;
+import domain.board.Square;
 
 class SquareTest {
     @Test
