+++ src/test/java/chess/repository/TestConnector.java
@@ -1,4 +1,4 @@
-package repository;
+package chess.repository;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
